ex5

+++ src/exercise3.c
@@ -4,13 +4,13 @@
 #define N 10
 
 int main(int argc, char** argv) {
-    int array[N];
-    int new_arr[N];
+    float array[N];
+    float new_arr[N];
     int sdvig = atoi(argv[1]);
 
     printf("Enter 10 numbers: ");
     for (int i = 0; i < N; ++i) {
-        scanf("%d", &array[i]);
+        scanf("%f", &array[i]);
     }
 
     for (int i = 0; i < N; ++i) {
@@ -19,7 +19,7 @@ int main(int argc, char** argv) {
 
     printf("New array:\n");
     for (int i = 0; i < N; ++i) {
-        printf("%d ", new_arr[i]);
+        printf("%f ", new_arr[i]);
     }
 
     return 0;
